fix: 내 정보 조회 response 타입 오류 수정

+++ Projects/Data/Sources/DTO/My/MyUserInfoDTO.swift
@@ -21,7 +21,7 @@ public struct MyUserInfoResponseModel: Decodable{
     public let nickname: String?
     public let level: Int?
     public let totalDistance: Double?
-    public let distanceToNextLevel: Int?
+    public let distanceToNextLevel: Double?
     public let totalKcal: Double?
     public let profileImageUrl: String?
     
@@ -30,8 +30,8 @@ public struct MyUserInfoResponseModel: Decodable{
                                nickname: nickname ?? "",
                                level: level ?? 0,
                                totalDistance: totalDistance ?? 0.0,
-                               distanceToNextLevel: distanceToNextLevel ?? 0,
-                               totalKcal: totalKcal ?? 0,
+                               distanceToNextLevel: distanceToNextLevel ?? 0.0,
+                               totalKcal: totalKcal ?? 0.0,
                                profileImageUrl: profileImageUrl)
     }
 }
