+++ Projects/Domain/Sources/Model/My/MyUserInfoModel.swift
@@ -12,11 +12,11 @@ public struct MyUserInfoModel{
     public let nickname: String
     public let level: Int
     public let totalDistance: Double
-    public let distanceToNextLevel: Int
+    public let distanceToNextLevel: Double
     public let totalKcal: Double
     public let profileImageUrl: String?
     
-    public init(userId: Int, nickname: String, level: Int, totalDistance: Double, distanceToNextLevel: Int, totalKcal: Double, profileImageUrl: String?) {
+    public init(userId: Int, nickname: String, level: Int, totalDistance: Double, distanceToNextLevel: Double, totalKcal: Double, profileImageUrl: String?) {
         self.userId = userId
         self.nickname = nickname
         self.level = level
